backend/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
httpx==0.25.2
python-multipart==0.0.6
python-dotenv==1.0.0